refactor: broke up search box inputs etc to improve style readability. fix: pagination items force entire page scroll rather than sitting inside their container. refactor: pagination controls moved to top on small devices
