+++ src/components/SearchResult/SearchResultList.js
@@ -17,14 +17,12 @@
  */
 import {useState, useEffect} from 'react';
 import Container from 'react-bootstrap/Container';
-import Form from 'react-bootstrap/Form';
-import Button from 'react-bootstrap/Button';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import SearchResultItem from './SearchResultItem';
 import LoadingOverlay from 'react-loading-overlay';
-import Pagination from 'react-bootstrap/Pagination';
 import SearchBox from './SearchBox';
+import PaginationBar from './PaginationBar';
 
 
 /**
@@ -53,7 +51,7 @@ const SearchResultList = ({
         }
         let newPageCount = Math.ceil(searchResults.hitCount / searchPageParams.size);
         setPageCount(newPageCount);
-    }, [searchResults, searchPageParams.size])
+    }, [searchResults, searchPageParams.size, searchPageParams.from])
 
     useEffect(() => {
         const from = (currentPageIndex-1) * searchPageParams.size;
@@ -65,18 +63,6 @@ const SearchResultList = ({
         if(pageNumber > 0) {
             setCurrentPageIndex(pageNumber)
         };
-    }
-
-    // prevent default to e.g. prevent page reload
-    // do NOT trigger search here, as the SearchBox
-    // component already triggers this by updating 
-    // the searchParams state.
-    const submit = (event) => {
-        event.preventDefault();
-    }
-
-    const toggleFilters = () => {
-        setShowFilters(!showFilters)
     }
 
     /**
@@ -92,28 +78,6 @@ const SearchResultList = ({
             }
             setSearchPageParams({...searchPageParams, size: e.target.value})
         } 
-    } 
-
-    const showSearchHelp = () => {
-        window.open(`${process.env.REACT_APP_BASE_URL}/SearchHelp_en.html`, '_blank');
-    }
-
-    const renderPaginationItems = () => {
-        let items = [];
-        // Calculate first index to render. This depends on the current page index as well as the
-        // total page count (which might be greater than the maximum number of buttons: 10).
-        let pagesToRender =  Math.min(9, pageCount - 1);
-        let firstIndex = Math.max(1, currentPageIndex - pagesToRender);
-        let lastIndex = firstIndex + pagesToRender;
-        for(let i = firstIndex; i <= lastIndex; i++){
-            items.push(<Pagination.Item 
-                key={i} 
-                active={i === currentPageIndex}
-                onClick={() => goToPage(i)}>
-                {i}
-            </Pagination.Item>)
-        }
-        return items;
     }
     
     const renderedSearchResults = searchResults.logs.length === 0 ? "No search results" : searchResults.logs.map((item, index) => {
@@ -126,82 +90,30 @@ const SearchResultList = ({
     });
 
     return(
-        <div className="grid-item full-height" style={{paddingLeft: "5px", paddingRight: "5px"}} >
-            <Form style={{paddingTop: "5px"}} onSubmit={(e) => submit(e)}>
-                <Form.Row>
-                    <Col style={{flexGrow: "0"}}>
-                        <Button size="sm" onClick={() => toggleFilters()}>{showFilters ? ">" : "<"}</Button>
-                    </Col>
-                    <Col style={{paddingLeft: "0px"}}>
-                    <SearchBox 
-                        {...{searchParams, setSearchParams, showFilters}}
-                    />
-                    </Col>
-                    <Col style={{flexGrow: "0" }}>
-                        <Button 
-                            size="sm"
-                            onClick={(e) => showSearchHelp()}>
-                            Help
-                        </Button>
-                    </Col>
-                </Form.Row>
-            </Form>
-            <LoadingOverlay
-                active={searchInProgress}
-                spinner
-                styles={{
-                    overlay: (base) => ({
-                        ...base,
-                        background: 'rgba(97, 97, 97, 0.3)',
-                        '& svg circle': {stroke: 'rgba(19, 68, 83, 0.9) !important'}
-                    })
-                    }}>
-
-                <div style={{overflowY: 'scroll', height: 'calc(80vh)'}}>
-                    {renderedSearchResults}
-                </div>
-                <div className="pagination-container">
-                    <Container>
-                        <Row>
-                            <Col style={{padding: '0px', maxWidth: '90px'}}>
-                            <Form.Label >Hits per page: </Form.Label> 
-                            </Col>
-                            <Col style={{padding: '0px', maxWidth: '50px'}}>
-                            <Form.Control size="sm" 
-                                type="input"
-                                value={searchPageParams.size}
-                                onChange={(e) => setPageSize(e)}
-                                />
-                            </Col>
-                        </Row>
-                        <Row style={{visibility: pageCount < 2 ? 'hidden' : 'visible'}}>
-                            <Col style={{marginTop: '13px', padding: '0px'}}>
-                            <Pagination
-                                size='sm'> 
-                                <Pagination.First disabled={currentPageIndex === 1}
-                                    onClick={() => goToPage(1)}
-                                    style={{fontWeight: 'bold'}}>&#124;&lt;</Pagination.First>
-                                <Pagination.Prev  onClick={() => goToPage(currentPageIndex - 1)}
-                                    disabled={currentPageIndex === 1}
-                                    style={{fontWeight: 'bold'}} >&lt;</Pagination.Prev>
-                                {renderPaginationItems()}
-                                <Pagination.Next onClick={() => goToPage(currentPageIndex + 1)} 
-                                    disabled={currentPageIndex === pageCount}
-                                    style={{fontWeight: 'bold'}}>&gt;</Pagination.Next>
-                                <Pagination.Last disabled={currentPageIndex === pageCount}
-                                    onClick={() => goToPage(pageCount)}
-                                    style={{fontWeight: 'bold'}}>&gt;&#124;</Pagination.Last>
-                                </Pagination>
-                            </Col>
-                            <Col style={{marginTop: '16px',  padding:'5px', maxWidth: '60px'}}>
-                                {currentPageIndex} / {pageCount}
-                            </Col>
-                        </Row>
-                    </Container>
-                </div>
-                
-            </LoadingOverlay>
-        </div>
+        <Container fluid className="grid-item h-100 p-0" >
+            <Row noGutters className='h-100 flex-column'>
+                <Col sm='auto' >
+                    <SearchBox {...{searchParams, setSearchParams, showFilters, setShowFilters}} />
+                </Col>    
+                <Col xs='auto' lg={{span: null, order: 3}}>
+                    <PaginationBar {...{pageCount, currentPageIndex, goToPage, searchPageParams, setPageSize}} />
+                </Col>
+                <Col style={{overflowY: 'scroll'}} className="border-top border-bottom" >
+                    <LoadingOverlay
+                    active={searchInProgress}
+                    spinner
+                    styles={{
+                        overlay: (base) => ({
+                            ...base,
+                            background: 'rgba(97, 97, 97, 0.3)',
+                            '& svg circle': {stroke: 'rgba(19, 68, 83, 0.9) !important'}
+                        })
+                        }}>
+                            {renderedSearchResults}
+                        </LoadingOverlay>
+                </Col>
+            </Row>
+        </Container>
     )
 }
